+++ pkg/lsx/attribute_test.go
@@ -1,0 +1,9 @@
+package lsx_test
+
+import (
+	"testing"
+)
+
+func TestAttributes(t *testing.T) {
+	t.Skip()
+}
