+++ pkg/lsx/meta_test.go
@@ -1,0 +1,9 @@
+package lsx_test
+
+import (
+	"testing"
+)
+
+func TestMeta(t *testing.T) {
+	t.Skip()
+}
