+++ pkg/lsx/node_test.go
@@ -1,0 +1,9 @@
+package lsx_test
+
+import (
+	"testing"
+)
+
+func TestNode(t *testing.T) {
+	t.Skip()
+}
