+++ pkg/lsx/reader_test.go
@@ -1,0 +1,9 @@
+package lsx_test
+
+import (
+	"testing"
+)
+
+func TestReader(t *testing.T) {
+	t.Skip()
+}
