+++ pkg/lsx/region_test.go
@@ -1,0 +1,9 @@
+package lsx_test
+
+import (
+	"testing"
+)
+
+func TestRegion(t *testing.T) {
+	t.Skip()
+}
