+++ pkg/lsx/version_test.go
@@ -1,0 +1,9 @@
+package lsx_test
+
+import (
+	"testing"
+)
+
+func TestVersion(t *testing.T) {
+	t.Skip()
+}
