+++ pkg/pak/file_reader_test.go
@@ -1,0 +1,9 @@
+package pak_test
+
+import (
+	"testing"
+)
+
+func TestFileReader(t *testing.T) {
+	t.Skip()
+}
