+++ pkg/pak/file_test.go
@@ -1,0 +1,9 @@
+package pak_test
+
+import (
+	"testing"
+)
+
+func TestReadFiles(t *testing.T) {
+	t.Skip()
+}
