+++ pkg/pak/header_test.go
@@ -1,0 +1,9 @@
+package pak_test
+
+import (
+	"testing"
+)
+
+func TestHeaderReader(t *testing.T) {
+	t.Skip()
+}
