+++ test/files/modsettings.lsx
